Fix g++ runtime compatibility. Fix tests/Makefile.

+++ runtime/ocamlcc-byterun-4.00/signals.c
@@ -244,7 +244,8 @@ static int posix_signals[] = {
 
 CAMLexport int caml_convert_signal_number(int signo)
 {
-  if (signo < 0 && signo >= -(sizeof(posix_signals) / sizeof(int)))
+  /* OCamlCC: fix g++ warning */
+  if (signo < 0 && signo >= -((int) (sizeof(posix_signals) / sizeof(int))))
     return posix_signals[-signo-1];
   else
     return signo;
@@ -253,7 +254,8 @@ CAMLexport int caml_convert_signal_number(int signo)
 CAMLexport int caml_rev_convert_signal_number(int signo)
 {
   int i;
-  for (i = 0; i < sizeof(posix_signals) / sizeof(int); i++)
+  /* OCamlCC: fix g++ warning */
+  for (i = 0; i < (int) (sizeof(posix_signals) / sizeof(int)); i++)
     if (signo == posix_signals[i]) return -i - 1;
   return signo;
 }
